Vladimir code was added.

git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@1001 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocclient/app/reper_app/gologramma/cuboid.cpp
@@ -597,7 +597,7 @@ std::vector<unsigned char> createImage(const std::vector<unsigned char> &imageMa
 
 QImage imageTransform(const std::vector<unsigned char>& in_image,const std::string& fileName)
 {
-    int size = static_cast<int>(sqrt(in_image.size()));
+    int size = static_cast<int>(sqrt(static_cast<double>(in_image.size())));
 
     QImage image(size,size,QImage::Format_RGB32);
